[Pdm2-4] Extract LoginForm from LoginScreen
